# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Safalife API is a comprehensive backend service for Safalife's competition management system. This API provides endpoints for health monitoring, user management, competition tracking, and administrative functions. Built with Go using Clean Architecture principles, it offers high performance and reliability for managing large-scale competitions.",
        "title": "Safalife API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Safalife API Support Team",
            "url": "http://www.safalife.com/support",
            "email": "api-support@safalife.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
...
        "/quran/audio/ayahs/{id}": {
            "get": {
                "description": "Get an ayah audio file by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Audio"
                ],
                "summary": "Get ayah audio file by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ayah audio file ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ayah audio file retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/quran.AyahAudioFileSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing ayah audio file",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Audio"
                ],
                "summary": "Update an ayah audio file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ayah audio file ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update ayah audio file request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateAyahAudioFileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ayah audio file updated successfully",
                        "schema": {
                            "$ref": "#/definitions/quran.UpdateAyahAudioFileSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an ayah audio file by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Audio"
                ],
                "summary": "Delete an ayah audio file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ayah audio file ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ayah audio file deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/quran.DeleteAyahAudioFileSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/quran/ayahs/{id}": {
            "get": {
                "description": "Get an ayah by its ID from the Quran",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ayahs"
                ],
                "summary": "Get ayah by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ayah ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "surah"
                        ],
                        "type": "string",
                        "example": "surah",
                        "description": "Include related data",
                        "name": "include",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ayah retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/quran.GetAyahByIdSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Ayah not found",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing ayah in the Quran",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ayahs"
                ],
                "summary": "Update an ayah",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ayah ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update ayah request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateAyahRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ayah updated successfully",
                        "schema": {
                            "$ref": "#/definitions/quran.UpdateAyahSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Ayah not found",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an existing ayah from the Quran",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ayahs"
                ],
                "summary": "Delete an ayah",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ayah ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ayah deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/quran.DeleteAyahSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Ayah not found",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "core.ErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                }
            }
        },
...
        "dto.AyahAudioFileResponse": {
            "type": "object",
            "properties": {
                "ayahId": {
                    "type": "string"
                },
                "byteSize": {
                    "type": "number"
                },
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "string"
                },
                "duration": {
                    "type": "number"
                },
                "filePath": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "reciterId": {
                    "type": "string"
                },
                "surahId": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "updatedBy": {
                    "type": "string"
                }
            }
        },
...
        "dto.AyahResponse": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "string"
                },
                "hizbNumber": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "juzNumber": {
                    "type": "integer"
                },
                "manzilNumber": {
                    "type": "integer"
                },
                "pageNumber": {
                    "type": "integer"
                },
                "surahId": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "updatedBy": {
                    "type": "string"
                }
            }
        },
...
        "dto.UpdateAyahAudioFileRequest": {
            "type": "object",
            "required": [
                "byteSize",
                "duration",
                "filePath",
                "id",
                "updatedBy"
            ],
            "properties": {
                "byteSize": {
                    "type": "number",
                    "minimum": 0
                },
                "duration": {
                    "type": "number",
                    "minimum": 0
                },
                "filePath": {
                    "type": "string",
                    "minLength": 1
                },
                "id": {
                    "type": "string"
                },
                "updatedBy": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                }
            }
        },
        "dto.UpdateAyahRequest": {
            "type": "object",
            "required": [
                "hizbNumber",
                "id",
                "juzNumber",
                "manzilNumber",
                "pageNumber",
                "text",
                "updatedBy"
            ],
            "properties": {
                "hizbNumber": {
                    "type": "integer",
                    "maximum": 60,
                    "minimum": 1
                },
                "id": {
                    "type": "string"
                },
                "juzNumber": {
                    "type": "integer",
                    "maximum": 30,
                    "minimum": 1
                },
                "manzilNumber": {
                    "type": "integer",
                    "maximum": 7,
                    "minimum": 1
                },
                "pageNumber": {
                    "type": "integer",
                    "maximum": 604,
                    "minimum": 1
                },
                "text": {
                    "type": "string",
                    "minLength": 1
                },
                "updatedBy": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                }
            }
        },
...
        "quran.AyahAudioFileSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.AyahAudioFileResponse"
                },
                "message": {
                    "type": "string",
                    "example": "Ayah audio file retrieved successfully"
                },
                "responseTime": {
                    "type": "string",
                    "example": "15.234ms"
                },
                "status_code": {
                    "type": "integer",
                    "example": 200
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "timestamp": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                }
            }
        },
...
        "quran.DeleteAyahAudioFileSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string",
                    "example": "null"
                },
                "message": {
                    "type": "string",
                    "example": "Ayah audio file deleted successfully"
                },
                "responseTime": {
                    "type": "string",
                    "example": "15.234ms"
                },
                "status_code": {
                    "type": "integer",
                    "example": 200
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "timestamp": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                }
            }
        },
        "quran.DeleteAyahSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string",
                    "example": "null"
                },
                "message": {
                    "type": "string",
                    "example": "Ayah deleted successfully"
                },
                "responseTime": {
                    "type": "string",
                    "example": "15.234ms"
                },
                "status_code": {
                    "type": "integer",
                    "example": 200
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "timestamp": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                }
            }
        },
...
        "quran.GetAyahByIdSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.AyahResponse"
                },
                "message": {
                    "type": "string",
                    "example": "Ayah retrieved successfully"
                },
                "responseTime": {
                    "type": "string",
                    "example": "15.234ms"
                },
                "status_code": {
                    "type": "integer",
                    "example": 200
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "timestamp": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                }
            }
        },
...
        "quran.QuranErrorResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "$ref": "#/definitions/core.ErrorDetail"
                },
                "message": {
                    "type": "string",
                    "example": "Failed to process request"
                },
                "responseTime": {
                    "type": "string",
                    "example": "15.234ms"
                },
                "status_code": {
                    "type": "integer",
                    "example": 500
                },
                "success": {
                    "type": "boolean",
                    "example": false
                },
                "timestamp": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                }
            }
        },
...
        "quran.UpdateAyahAudioFileSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.AyahAudioFileResponse"
                },
                "message": {
                    "type": "string",
                    "example": "Ayah audio file updated successfully"
                },
                "responseTime": {
                    "type": "string",
                    "example": "15.234ms"
                },
                "status_code": {
                    "type": "integer",
                    "example": 200
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "timestamp": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                }
            }
        },
        "quran.UpdateAyahSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.AyahResponse"
                },
                "message": {
                    "type": "string",
                    "example": "Ayah updated successfully"
                },
                "responseTime": {
                    "type": "string",
                    "example": "15.234ms"
                },
                "status_code": {
                    "type": "integer",
                    "example": 200
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "timestamp": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Enter the token with the `Bearer: ` prefix, e.g. \"Bearer abcde12345\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}